Fix dancing download notifications on phones.
There's always going to be some amount of visual noise when
notifications change order, and this can happen any time a
notification changes its "when" field (as is the case for
download progress bars). But in HC we added an optimization
that will reuse the current row if the download is the last
item in the list (so a fresher "when" will still put it at
the end), and that optimization is re-applied here.

Now, if you have multiple notifications all competing for
the top spot, I can't help you. But we had that problem
before (it just wasn't as egregious because we didn't have
the shiny animations).

Bug: 4971404
Change-Id: I5bcc905bdb19a8eb5b4494a6ba9825b4d5373caa