Add IS_READ field to CallLog and VoicemailContract.
We had initially decided to merge the NEW field (from the call log) and
IS_READ field (from the voicemail contract) into the single NEW field
already present in the call log.

However, it turned out that the meaning of the two fields is slightly
different: NEW means the items has just been inserted (and the user is
not aware of it) while IS_READ represents whether the user has read this
specific message (or heard, in case of voicemails).

This change makes the IS_READ field public, as well as deletes the NEW
field from the voicemail contract, since it is no longer needed there.
The NEW field in the voicemail contract (in fact the entire voicemail
contract) has never been released (this feature is new in ICS).

Bug: 5036195
Change-Id: I740c51c1a8e6d2460050eaab9943fb38b1565058