Fix issue #5016544: IME keyboard hides folder rename text field.
Tweak some issues in TextView with the focus rect used to determing
where the scroll position needs to be.  The ultimate problem was that
in various situations it would use the right-most selection position
cursor as the scroll location and when it adds 1 to make this a valid
rect we end up with a rectangle that is outside of the view.

Change-Id: Ia200c58e4e014d2a0a1be4761f9a1e5eb702a8e5