Relax form autocomplete conditions
Use the URL host and path rather than the complete url to store
form autocomplete data. This helps in the situation that a site
uses some dynamic query string on the page that contains the form.

Also set the autoocmplete threshold to 1 so that we don't flick the
autocomplete options up and down as you type the first few characters.

Bug: 5265606
Change-Id: I7b372400062ae34f70a78b786007910dc179b101