Remove isAnyActiveDataConnection and promote isDisconnected.
As Yoonsung Nam pointed out in the bug using isDisconnected
is more precise here and this was the only place
isAnyActiveDataConnection was used. So I've removed it
and promoted isDisconnected to DataConnectionTracker.

It should also be noted that although the Framework should
be as graceful as possible, it is still a requirement that
the radio must be turned off/on anytime the Framework
issues the appropriate command.

Bug: 5306201
Change-Id: I95130451e11c51e43406b57eb538d01f8dde61be