AAH: Fix PID gain for HW clock recovery.
Before HW came in, when doing SW clock recovery, gain had been set to
produce a correction expressed in PPM from [-100, 100] PPM.  The HW
local clock HAL has been spec'ed to take a drive range from MIN_INT_16
to MAX_INT_16.  This changes the PID recovery loop to use a range from
MIN_INT_16 to MAX_INT_16, and adjusts the P and I gains to be the same
as they were before, but adjusted for the scale change.

Change-Id: Ia636fd1bc6c5c1117ddfdc5289c51286cb874c23